refactor(core): :recycle: rename branding-preference file to get-branding-preference
